+++ src/frontend/apps/desk/src/core/config/types.ts
@@ -1,4 +1,5 @@
 export interface Config {
+  CRISP_WEBSITE_ID: string;
   LANGUAGES: [string, string][];
   RELEASE: string;
   COMMIT: string;